refactor project and student search bars into one
